collect common part and extend TestCase class
